+++ app/Http/Controllers/web/admin/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\web\admin;
 use App\Http\Controllers\Controller;
 use App\Models\Permission;
 use App\Models\Role;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -12,21 +13,25 @@ class RoleController extends Controller
 {
     public function index()
     {
-        $roles = Role::latest()->paginate(10);
+        $this->authorize('viewIndex' , Role::class);
+        $roles = Role::whereNot('name' , 'user')->whereNot('name' , 'admin')->latest()->paginate(10);
         return view('admin.roles.index', compact('roles'));
     }
 
     public function create()
     {
+        $this->authorize('create' , Role::class);
         $permissions = Permission::all();
         return view('admin.roles.create', compact('permissions'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create' , Role::class);
+
         $request->validate([
             'name' => 'required',
-            'permissions' => 'required'
+            'permissions' => 'required',
         ]);
 
         try {
@@ -34,28 +39,30 @@ class RoleController extends Controller
 
             $role = Role::create([
                 'name' => $request->name,
-                'guard_name' => 'web'
             ]);
-            $role->givePermissionTo($request->permissions);
+
+            $role->permissions()->attach($request->permissions);
 
             DB::commit();
         } catch (\Exception $ex) {
             DB::rollBack();
-            dd($ex->getMessage());
-            return redirect()->back();
+            return redirect()->back()->with('message' , 'نقش مور نظر ثبت نشد.');
         }
         return redirect()->route('admin.roles.index');
     }
 
     public function show(Role $role)
     {
+        $this->authorize('viewShow' , Role::class);
         return view('admin.roles.show', compact('role'));
     }
 
     public function edit(Role $role)
     {
+        $this->authorize('edit' , Role::class);
+
         if($role->name == 'user' || $role->name == 'admin')
-            return back();
+            return back()->with('message' , 'شما نمی توانید این کار را انجام دهید.');
 
         $permissions = Permission::all();
         return view('admin.roles.edit', compact('role', 'permissions'));
@@ -63,8 +70,10 @@ class RoleController extends Controller
 
     public function update(Request $request, Role $role)
     {
+        $this->authorize('edit' , Role::class);
+
         if($role->name == 'user' || $role->name == 'admin')
-            return back();
+            return back()->with('message' , 'شما نمی توانید این کار را انجام دهید.');
 
         $request->validate([
             'name' => 'required',
@@ -77,11 +86,13 @@ class RoleController extends Controller
             $role->update([
                 'name' => $request->name,
             ]);
-            $role->syncPermissions($request->permissions);
+
+            $role->permissions()->sync($request->permissions);
 
             DB::commit();
         } catch (\Exception $ex) {
             DB::rollBack();
+
             return redirect()->back();
         }
         return redirect()->route('admin.roles.index');
@@ -89,8 +100,10 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        $this->authorize('delete' , Role::class);
+
         if($role->name == 'user' || $role->name == 'admin')
-            return back();
+            return back()->with('message' , 'شما نمی توانید این کار را انجام دهید.');
 
         $role->delete();
         return back();
